Deleted ShadowRaytracerSRP and modfied ShadowRaytracer.cs
