--- Editor/ScriptsSRP.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8c5ffcc798ce70943b33707a099d1b90
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
