--- Editor/ScriptsSRP/ExportToImageWindowSRP.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e52b2567bcf942e4598ce46c9ef30259
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
