--- Editor/ScriptsSRP/ShadowRaytracerSRP_Editor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8589b905824a0fc46b49534452816666
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
