--- Runtime/ScriptsSRP.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7c0fe2f4e88d51743b892f3655fc2daf
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
