--- Runtime/ScriptsSRP/ShadowRaytracerSRP.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9977555669b5a0a4ebbc48ed3acdcca9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
