+++ SamplesUniversalRP~/PrimitiveObjects/PrimitiveObjectsScene.unity
@@ -98,7 +98,7 @@ LightmapSettings:
     m_TrainingDataDestination: TrainingData
     m_LightProbeSampleCountMultiplier: 4
   m_LightingDataAsset: {fileID: 0}
-  m_UseShadowmask: 0
+  m_UseShadowmask: 1
 --- !u!196 &4
 NavMeshSettings:
   serializedVersion: 2
@@ -727,7 +727,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 963194225}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 9977555669b5a0a4ebbc48ed3acdcca9, type: 3}
+  m_Script: {fileID: 11500000, guid: b21f82ee99238ee4eb21b9a59244893c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   m_generateRenderTexture: 0
@@ -760,7 +760,6 @@ MonoBehaviour:
   m_antialiasing: 0
   m_clampBlendshapeWeights: 0
   m_dbgVerboseLog: 0
-  m_DirectionalLightList: []
   m_foldDebug: 0
 --- !u!114 &963194230
 MonoBehaviour:
